split functions, clean vault on install, use time limit only, webhook cleanup
