+++ apps/supabase/src/__tests__/bundle.test.ts
@@ -6,19 +6,39 @@ import { beforeAll, describe, expect, it } from 'vitest'
 // ---------------------------------------------------------------------------
 
 describe.concurrent('Bundled edge function code', () => {
+  let setupCode: string
+  let webhookCode: string
   let syncCode: string
 
   beforeAll(async () => {
     const distPath = resolve(import.meta.dirname, '../../dist/index.js')
     const mod = await import(distPath)
+    setupCode = mod.setupFunctionCode
+    webhookCode = mod.webhookFunctionCode
     syncCode = mod.syncFunctionCode
   })
 
-  it('consolidated edge function is bundled', () => {
+  it('setup edge function is bundled', () => {
+    expect(setupCode).toBeTruthy()
+  })
+
+  it('webhook edge function is bundled', () => {
+    expect(webhookCode).toBeTruthy()
+  })
+
+  it('sync edge function is bundled', () => {
     expect(syncCode).toBeTruthy()
   })
 
-  it('bundled code contains Deno.serve entry point', () => {
+  it('setup code contains Deno.serve entry point', () => {
+    expect(setupCode).toContain('Deno.serve')
+  })
+
+  it('webhook code contains Deno.serve entry point', () => {
+    expect(webhookCode).toContain('Deno.serve')
+  })
+
+  it('sync code contains Deno.serve entry point', () => {
     expect(syncCode).toContain('Deno.serve')
   })
