+++ apps/supabase/src/supabase.ts
@@ -1,5 +1,5 @@
 import { SupabaseManagementAPI } from 'supabase-management-js'
-import { syncFunctionCode } from './edge-function-code.js'
+import { setupFunctionCode, webhookFunctionCode, syncFunctionCode } from './edge-function-code.js'
 import pkg from '../package.json' with { type: 'json' }
 import { parseSchemaComment, StripeSchemaComment } from './schemaComment.js'
 
@@ -160,7 +160,7 @@ export class SupabaseSetupClient {
         '${schedule}',
         $$
         SELECT net.http_post(
-          url := 'https://${this.projectRef}.${this.projectBaseUrl}/functions/v1/stripe-worker/sync',
+          url := 'https://${this.projectRef}.${this.projectBaseUrl}/functions/v1/stripe-worker',
           headers := jsonb_build_object(
             'Authorization', 'Bearer ' || (SELECT decrypted_secret FROM vault.decrypted_secrets WHERE name = 'stripe_sync_worker_secret')
           )
@@ -180,7 +180,7 @@ export class SupabaseSetupClient {
    * Get the webhook URL for this project
    */
   getWebhookUrl(): string {
-    return `https://${this.projectRef}.${this.projectBaseUrl}/functions/v1/stripe-worker/webhook`
+    return `https://${this.projectRef}.${this.projectBaseUrl}/functions/v1/stripe-webhook`
   }
 
   /**
@@ -387,12 +387,7 @@ export class SupabaseSetupClient {
         await this.updateComment({ status: 'uninstalling', startTime })
       }
 
-      // Invoke the DELETE endpoint on the consolidated stripe-sync function
-      const setupResult = await this.invokeFunction(
-        'stripe-worker/setup',
-        'DELETE',
-        this.accessToken
-      )
+      const setupResult = await this.invokeFunction('stripe-setup', 'DELETE', this.accessToken)
 
       if (!setupResult.success) {
         throw new Error(`Uninstall failed: ${setupResult.error}`)
@@ -457,12 +452,16 @@ export class SupabaseSetupClient {
       }
       await this.setSecrets(secrets)
 
-      // Deploy the single consolidated edge function
+      // Deploy edge functions
+      const versionedSetup = this.injectPackageVersion(setupFunctionCode, version)
+      const versionedWebhook = this.injectPackageVersion(webhookFunctionCode, version)
       const versionedSync = this.injectPackageVersion(syncFunctionCode, version)
+      await this.deployFunction('stripe-setup', versionedSetup, false)
+      await this.deployFunction('stripe-webhook', versionedWebhook, false)
       await this.deployFunction('stripe-worker', versionedSync, false)
 
-      // Run setup (migrations + webhook creation) via the /setup path
-      const setupResult = await this.invokeFunction('stripe-worker/setup', 'POST', this.accessToken)
+      // Run setup (migrations + webhook creation)
+      const setupResult = await this.invokeFunction('stripe-setup', 'POST', this.accessToken)
 
       if (!setupResult.success) {
         throw new Error(`Setup failed: ${setupResult.error}`)
@@ -482,10 +481,10 @@ export class SupabaseSetupClient {
       // if done before.
       if (!skipInitialSync) {
         try {
-          await this.invokeFunction('stripe-worker/sync', 'POST', this.workerSecret)
+          await this.invokeFunction('stripe-worker', 'POST', this.workerSecret)
         } catch (error) {
           const errorMessage = error instanceof Error ? error.message : String(error)
-          console.warn(`Failed to invoke stripe-worker/sync: ${errorMessage}`)
+          console.warn(`Failed to invoke stripe-worker: ${errorMessage}`)
         }
       }
     } catch (error) {
